requirements.txt
tornado==3.1.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
facebook-sdk==0.4.
itsdangerous==0.23
mutagen==1.22
pymongo==2.6.3